task(refactor): Refactors Checkup and CheckupCLI classes into one
