+++ packages/cli/src/index.ts
@@ -1,8 +1,7 @@
 import { Command, flags } from '@oclif/command';
+import { TaskConstructor, TaskList, getRegisteredTasks, ui } from '@checkup/core';
 
-import Checkup from './checkup';
-
-class CheckupCli extends Command {
+class Checkup extends Command {
   static description = 'A CLI that provides health check information about your project';
 
   static flags = {
@@ -15,14 +14,47 @@ class CheckupCli extends Command {
   };
 
   async run() {
-    let { args, flags } = this.parse(CheckupCli);
+    let { flags } = this.parse(Checkup);
+    let registeredTasks: TaskConstructor[];
 
     await this.config.runHook('register-tasks', {});
 
-    let checkup = new Checkup(args, flags);
+    registeredTasks = getRegisteredTasks();
 
-    return checkup.run();
+    let tasksToBeRun = new TaskList();
+
+    if (flags.task !== undefined) {
+      let task = Object.values(registeredTasks).find(
+        task => flags.task === task.name.replace('Task', '')
+      );
+
+      if (task !== undefined) {
+        tasksToBeRun.addTask(task);
+      }
+    } else {
+      tasksToBeRun.addTasks(registeredTasks);
+    }
+
+    ui.action.start('Checking up on your project');
+    let taskResults = await tasksToBeRun.runTasks();
+
+    if (!flags.silent) {
+      if (flags.json) {
+        let resultData = {};
+        taskResults.forEach(taskResult => {
+          resultData = Object.assign(resultData, taskResult.toJson());
+        });
+
+        ui.styledJSON(resultData);
+      } else {
+        taskResults.forEach(result => {
+          result.toConsole();
+        });
+      }
+    }
+
+    ui.action.stop();
   }
 }
 
-export = CheckupCli;
+export = Checkup;
